Fix : update all dll funct to work with libft.h

+++ libft/src/dll_insert_head.c
@@ -6,14 +6,13 @@
 /*   By: lagea <lagea@student.s19.be>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 13:18:21 by lagea             #+#    #+#             */
-/*   Updated: 2024/05/20 18:33:44 by lagea            ###   ########.fr       */
+/*   Updated: 2024/05/24 17:01:44 by lagea            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/libft.h"
-#include "../../inc/push_swap.h"
 
-void	dll_insert_head(ssize_t data, t_stack *stack)
+void	dll_insert_head(int data, t_stack *stack)
 {
 	t_node	*new;
