+++ libft/src/dll_insert_tail.c
@@ -6,14 +6,13 @@
 /*   By: lagea <lagea@student.s19.be>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 13:34:55 by lagea             #+#    #+#             */
-/*   Updated: 2024/05/23 14:37:44 by lagea            ###   ########.fr       */
+/*   Updated: 2024/05/24 17:01:40 by lagea            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/libft.h"
-#include "../../inc/push_swap.h"
 
-void	dll_insert_tail(ssize_t data, t_stack *stack)
+void	dll_insert_tail(int data, t_stack *stack)
 {
 	t_node	*new;
